Revert "[GCU Feature Update] Cherry-pick Platform Validator PR to 202205  (#2883)"
This reverts commit 436ecb06f955409730fd6958a527b8c7685d7a42.
